Expose redactor
